third_party/blink/web_tests/external/wpt/dom/events/webkit-animation-iteration-event-expected.txt

This is a testharness.js-based test.
FAIL onanimationiteration and onwebkitanimationiteration are not aliases assert_equals: onwebkitanimationiteration should initially be null expected (object) null but got (undefined) undefined
FAIL dispatchEvent of a webkitAnimationIteration event does trigger a prefixed event handler or listener assert_equals: onwebkitanimationiteration must be supported and not previously set expected (object) null but got (undefined) undefined
PASS dispatchEvent of a webkitAnimationIteration event does not trigger an unprefixed event handler or listener
FAIL dispatchEvent of an animationiteration event does not trigger a prefixed event handler or listener assert_equals: onwebkitanimationiteration must be supported and not previously set expected (object) null but got (undefined) undefined
FAIL onwebkitanimationiteration event handler should trigger for an animation assert_equals: onwebkitanimationiteration must be supported and not previously set expected (object) null but got (undefined) undefined
FAIL onwebkitanimationiteration event handler should not trigger if an unprefixed event handler also exists assert_equals: onwebkitanimationiteration must be supported and not previously set expected (object) null but got (undefined) undefined
FAIL onwebkitanimationiteration event handler should not trigger if an unprefixed listener also exists assert_equals: onwebkitanimationiteration must be supported and not previously set expected (object) null but got (undefined) undefined
FAIL event types for prefixed and unprefixed animationiteration event handlers should be named appropriately assert_equals: onwebkitanimationiteration must be supported and not previously set expected (object) null but got (undefined) undefined
PASS webkitAnimationIteration event listener should trigger for an animation
PASS webkitAnimationIteration event listener should not trigger if an unprefixed listener also exists
PASS webkitAnimationIteration event listener should not trigger if an unprefixed event handler also exists
PASS event types for prefixed and unprefixed animationiteration event listeners should be named appropriately
PASS webkitAnimationIteration event listener is case sensitive
Harness: the test ran to completion.
